relocate parseGethGenesis tests
