+++ packages/common/tests/data/geth-genesis/invalid-spurious-dragon.json
