+++ packages/common/tests/data/geth-genesis/no-extra-data.json
