+++ packages/common/tests/data/geth-genesis/pow.json
