+++ packages/common/tests/data/geth-genesis/testnet.json
